fastapi_app/requirements.txt
fastapi
uvicorn
pydantic
pygithub
python-dotenv